лабораторная 1
